pyproject.toml
[build-system]
requires = [
    "setuptools>=61.0.0",
]
build-backend = "setuptools.build_meta"

[project]
name = "nomad-analysis"
version = '0.1.0'
description = "A standard NOMAD plugin for analysis."
readme = "README.md"
authors = [
    { name = "Sarthak Kapoor", email = "sarthak.kapoor@physik.hu-berlin.de" }
]
license = { text = "Apache-2.0" }
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "License :: OSI Approved :: Apache Software License",
]
dependencies = [
    "nomad-lab>=1.2.2dev457",
    "nbformat>=5.9.2",
]
[project.optional-dependencies]
dev = [
    "ruff>=0.2.0",
    "pytest",
    "structlog==22.3.0",
    "python-logstash",
]

[project.urls]
"Homepage" = "https://github.com/FAIRmat-NFDI/nomad-analysis"
"Bug Tracker" = "https://github.com/FAIRmat-NFDI/nomad-analysis/issues"
    
[tool.ruff]   
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4

# Assume Python 3.9
target-version = "py39"

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
   # isort
    "I",
    # pylint 
    "PL"
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
]
    
fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
nomad_analysis = ['*/nomad_plugin.yaml']
